requirements.txt
selenium==4.41.0
beautifulsoup4==4.14.3
python-dotenv==1.2.2
pandas==2.3.3
webdriver-manager==4.0.2
streamlit==1.55.0
plotly==6.6.0
browser-cookie3>=0.20.0
